+++ packages/integrations/src/index.ts
@@ -1,7 +1,8 @@
-import Namada from "./Namada";
 import Keplr from "./Keplr";
 import Metamask from "./Metamask";
+import Namada from "./Namada";
 
-export { Namada, Keplr, Metamask };
 export * from "./hooks";
 export type { MetamaskWindow } from "./Metamask";
+export * from "./utils";
+export { Keplr, Metamask, Namada };
